+++ c/spot/source/t13supp.c
@@ -518,7 +518,7 @@ static void t13Read(cffCtx h, Offset offset, int init, int csLen) {
                         int iTop = h->stack.cnt - 1;
                         int iBottom = h->stack.cnt - n;
 
-                        if (n < 0 || iBottom < 0)
+                        if (n <= 0 || iBottom < 0)
                             fatal(h, "limit check (roll)");
 
                         /* Constrain j to [0,n) */